Moved drafting code from FragmentDrafter to ActivityDrafter * New activity added to AndroidManifest * removed TODO fromActivitySupply
